--- src/pages/post/postForm/Header.jsx
@@ -1,12 +1,0 @@
-import Icon from '@/components/common/icon/Icon';
-
-function Header() {
-	return (
-		<div>
-			<Icon name='logo' type='img' />
-			Rolling
-		</div>
-	);
-}
-
-export default Header;
