--- src/pages/post/postForm/postForm.jsx
@@ -1,14 +1,0 @@
-import '@/pages/post/postForm/postForm.scss';
-
-import Layout from '@/components/common/Layout';
-import SelectForm from '@/pages/post/postForm/SelectForm';
-
-function PostForm() {
-	return (
-		<Layout>
-			<SelectForm />
-		</Layout>
-	);
-}
-
-export default PostForm;
