--- src/pages/post/postForm/postForm.scss
@@ -1,41 +1,0 @@
-form {
-	width: 72rem;
-	display: flex;
-	flex-direction: column;
-	margin: 0 auto;
-}
-
-.inputBox {
-	display: flex;
-	flex-direction: column;
-	align-items: flex-start;
-	gap: 1.2rem;
-	margin-top: 5.7rem;
-}
-
-.sendTo {
-	@include typo_24_bold;
-	border: none;
-}
-
-.textBox {
-	display: flex;
-	flex-direction: column;
-	align-items: flex-start;
-	gap: 0.4rem;
-	margin-top: 5rem;
-}
-
-.textBox > h2 {
-	@include typo_24_bold;
-}
-
-.textBox > p {
-	@include typo_16_regular;
-}
-
-.toggleBox {
-	display: flex;
-	flex-direction: row;
-	margin: 2.4rem 0;
-}
